Helm: unify NGINX and gateway values sections (#3181)
Unifies the GEM gateway and NGINX deployment under the same section in the values file - gateway. The idea is to remove the nginx-specific values in release 7.0.0.

In doing so it also introduces the following changes:

nginx and the gateway now use the same Service
nginx and the gateway now use the same Deployment
nginx and the gateway now use the same Ingress
These three together make it seamless to migrate from Mimir to GEM.

By unifying the two the GEM gateway now has autoscaling and support for OpenShift Route.

The PR also add a migration guide from nginx to gateway.